refactoring - move devices-related stuff from tvm.runtime.ndarray to tvm.runtime.devices
