+++ python/tvm/__init__.py
@@ -31,9 +31,7 @@ from ._ffi import register_object, register_func, register_extension, get_global
 # top-level alias
 # tvm.runtime
 from .runtime.object import Object
-from .runtime.ndarray import device, cpu, cuda, gpu, opencl, cl, vulkan, metal, mtl
-from .runtime.ndarray import vpi, rocm, ext_dev, hexagon
-from .runtime import ndarray as nd
+from .runtime import ndarray as nd, device, cpu, cuda, gpu, rocm, opencl, metal, vpi, vulkan, ext_dev, hexagon, cl, mtl
 
 # tvm.error
 from . import error
