+++ python/tvm/relay/analysis/analysis.py
@@ -22,7 +22,7 @@ configuring the passes and scripting them in Python.
 """
 from ...ir import IRModule
 from ...relay import transform, build_module
-from ...runtime.ndarray import cpu
+from ...runtime.devices import cpu
 
 from . import _ffi_api
 from .feature import Feature
