+++ python/tvm/relay/testing/init.py
@@ -19,6 +19,7 @@ from functools import reduce
 import numpy as np
 
 import tvm
+import tvm.runtime
 from tvm import relay
 
 
@@ -176,5 +177,5 @@ def create_workload(net, initializer=None, seed=0):
             continue
         init_value = np.zeros(v.concrete_shape).astype(v.dtype)
         initializer(k, init_value)
-        params[k] = tvm.nd.array(init_value, device=tvm.cpu(0))
+        params[k] = tvm.nd.array(init_value, device=tvm.runtime.cpu(0))
     return mod, params
