+++ python/tvm/rpc/client.py
@@ -22,9 +22,9 @@ import struct
 import time
 
 import tvm._ffi
+import tvm.runtime
 from tvm._ffi.base import TVMError
 from tvm.contrib import utils
-from tvm.runtime import ndarray as nd
 
 from . import _ffi_api, base, server
 
@@ -84,7 +84,7 @@ class RPCSession(object):
         dev: Device
             The corresponding encoded remote device.
         """
-        dev = nd.device(dev_type, dev_id)
+        dev = tvm.runtime.device(dev_type, dev_id)
         encode = (self._tbl_index + 1) * base.RPC_SESS_MASK
         dev.device_type += encode
         dev._rpc_sess = self
