+++ python/tvm/topi/cuda/conv2d_alter_op.py
@@ -19,6 +19,7 @@
 
 import logging
 import tvm
+import tvm.runtime
 from tvm import te, relay, autotvm
 
 from .. import nn
@@ -231,7 +232,7 @@ def _alter_conv2d_layout(attrs, inputs, tinfos, out_type):
 
     if topi_tmpl == "conv2d_HWNCnc_tensorcore.cuda":
         assert data_layout == "HWNC" and kernel_layout == "HWOI"
-        assert float(tvm.cuda(0).compute_version) >= 7.5
+        assert float(tvm.runtime.cuda(0).compute_version) >= 7.5
         H, W, N, CI = get_const_tuple(data.shape)
         KH, KW, CO, _ = get_const_tuple(kernel.shape)
